+++ test/Remote.Linq.Tests/Serialization/Expressions/When_using_simple_bool_expression.cs
@@ -48,8 +48,8 @@ public abstract class When_using_simple_bool_expression
     [Fact]
     public void System_expresison_should_be_equal()
     {
-        var exp1 = _remoteExpression.ToLinqExpression<bool, bool>();
-        var exp2 = _serializedRemoteExpression.ToLinqExpression<bool, bool>();
+        var exp1 = _remoteExpression.ToLinqExpression<Func<bool, bool>>();
+        var exp2 = _serializedRemoteExpression.ToLinqExpression<Func<bool, bool>>();
 
         exp1.ShouldEqualExpression(exp2);
     }